Files are renamed and moved.
